+++ index.html
@@ -15,7 +15,7 @@
 		<aside>
 			<side-menu :nav='nav'></side-menu>
 		</aside>
-		<login-dialog :user="user" :servers="config.servers"></login-dialog>
+		<login :user="user" :servers="config.servers"></login>
 	</section>
 	<section id="oc-content" class="uk-offcanvas-content"></section>
 	<script data-main="core/js/core.bundle.js" src="node_modules/requirejs/require.js"></script>
